requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.1.10
django-webpack-loader==0.6.0
djangorestframework==3.10.1
et-xmlfile==1.0.1
isort==4.3.20
jdcal==1.4.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
openpyxl==2.6.2
Pillow==6.0.0
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
